+++ pkg/api/auth/local/session_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2024-Present The UDS Authors
 
-package auth
+package local
 
 import (
 	"net/http"
